+++ src/provinces/province.cpp
@@ -292,7 +292,7 @@ void restore_unsaved_values(sys::state& state) {
 
 	for(auto si : state.world.in_state_instance) {
 		province::for_each_province_in_state_instance(state, si, [&](dcon::province_id p) {
-			if(state.world.province_get_building_level(p, economy::province_building_type::naval_base) > 0) {
+			if(state.world.province_get_building_level(p, uint8_t(economy::province_building_type::naval_base)) > 0) {
 				state.world.state_instance_set_naval_base_is_taken(si, true);
 			} else {
 				for(auto pc : state.world.province_get_province_building_construction(p)) {
@@ -320,8 +320,8 @@ bool generic_can_build_railroads(sys::state& state, dcon::province_id id, dcon::
 	if(military::province_is_under_siege(state, id))
 		return false;
 
-	int32_t current_rails_lvl = state.world.province_get_building_level(id, economy::province_building_type::railroad);
-	int32_t max_local_rails_lvl = state.world.nation_get_max_building_level(n, economy::province_building_type::railroad);
+	int32_t current_rails_lvl = state.world.province_get_building_level(id, uint8_t(economy::province_building_type::railroad));
+	int32_t max_local_rails_lvl = state.world.nation_get_max_building_level(n, uint8_t(economy::province_building_type::railroad));
 	int32_t min_build_railroad =
 			int32_t(state.world.province_get_modifier_values(id, sys::provincial_mod_offsets::min_build_railroad));
 
@@ -353,8 +353,8 @@ bool can_build_railroads(sys::state& state, dcon::province_id id, dcon::nation_i
 			return false;
 	}
 
-	int32_t current_rails_lvl = state.world.province_get_building_level(id, economy::province_building_type::railroad);
-	int32_t max_local_rails_lvl = state.world.nation_get_max_building_level(n, economy::province_building_type::railroad);
+	int32_t current_rails_lvl = state.world.province_get_building_level(id, uint8_t(economy::province_building_type::railroad));
+	int32_t max_local_rails_lvl = state.world.nation_get_max_building_level(n, uint8_t(economy::province_building_type::railroad));
 	int32_t min_build_railroad =
 			int32_t(state.world.province_get_modifier_values(id, sys::provincial_mod_offsets::min_build_railroad));
 
@@ -375,8 +375,8 @@ bool can_build_fort(sys::state& state, dcon::province_id id, dcon::nation_id n) 
 	if(military::province_is_under_siege(state, id))
 		return false;
 
-	int32_t current_lvl = state.world.province_get_building_level(id, economy::province_building_type::fort);
-	int32_t max_local_lvl = state.world.nation_get_max_building_level(n, economy::province_building_type::fort);
+	int32_t current_lvl = state.world.province_get_building_level(id, uint8_t(economy::province_building_type::fort));
+	int32_t max_local_lvl = state.world.nation_get_max_building_level(n, uint8_t(economy::province_building_type::fort));
 	int32_t min_build = int32_t(state.world.province_get_modifier_values(id, sys::provincial_mod_offsets::min_build_fort));
 
 	return (max_local_lvl - current_lvl - min_build > 0) && !has_fort_being_built(state, id);
@@ -400,8 +400,8 @@ bool can_build_naval_base(sys::state& state, dcon::province_id id, dcon::nation_
 
 	auto si = state.world.province_get_state_membership(id);
 
-	int32_t current_lvl = state.world.province_get_building_level(id, economy::province_building_type::naval_base);
-	int32_t max_local_lvl = state.world.nation_get_max_building_level(n, economy::province_building_type::naval_base);
+	int32_t current_lvl = state.world.province_get_building_level(id, uint8_t(economy::province_building_type::naval_base));
+	int32_t max_local_lvl = state.world.nation_get_max_building_level(n, uint8_t(economy::province_building_type::naval_base));
 	int32_t min_build = int32_t(state.world.province_get_modifier_values(id, sys::provincial_mod_offsets::min_build_naval_base));
 
 	return (max_local_lvl - current_lvl - min_build > 0) && (current_lvl > 0 || !si.get_naval_base_is_taken()) && !has_naval_base_being_built(state, id);
@@ -443,8 +443,8 @@ bool can_build_province_building(sys::state& state, dcon::province_id id, dcon::
 			return false;
 	}
 
-	int32_t current_lvl = state.world.province_get_building_level(id, t);
-	int32_t max_local_lvl = state.world.nation_get_max_building_level(n, t);
+	int32_t current_lvl = state.world.province_get_building_level(id, uint8_t(t));
+	int32_t max_local_lvl = state.world.nation_get_max_building_level(n, uint8_t(t));
 	int32_t min_build = int32_t(state.world.province_get_modifier_values(id, sys::provincial_mod_offsets::min_build_railroad));
 	return (max_local_lvl - current_lvl - min_build > 0) && !has_railroads_being_built(state, id);
 }
@@ -685,7 +685,7 @@ void change_province_owner(sys::state& state, dcon::province_id id, dcon::nation
 		&& state.world.nation_get_is_civilized(new_owner) == true)
 		|| (!old_owner);
 	if(old_si) {
-		if(state.world.province_get_building_level(id, economy::province_building_type::naval_base) > 0) {
+		if(state.world.province_get_building_level(id, uint8_t(economy::province_building_type::naval_base)) > 0) {
 			state.world.state_instance_set_naval_base_is_taken(old_si, false);
 		} else {
 			for(auto pc : state.world.province_get_province_building_construction(id)) {
@@ -713,7 +713,7 @@ void change_province_owner(sys::state& state, dcon::province_id id, dcon::nation
 			state.world.province_set_is_slave(id, false);
 			if(will_be_colonial)
 				state.world.nation_set_is_colonial_nation(new_owner, true);
-			if(state.world.province_get_building_level(id, economy::province_building_type::naval_base) > 0)
+			if(state.world.province_get_building_level(id, uint8_t(economy::province_building_type::naval_base)) > 0)
 				state.world.state_instance_set_naval_base_is_taken(new_si, true);
 
 			state_is_new = true;
@@ -721,9 +721,9 @@ void change_province_owner(sys::state& state, dcon::province_id id, dcon::nation
 			auto sc = state.world.state_instance_get_capital(new_si);
 			state.world.province_set_is_colonial(id, state.world.province_get_is_colonial(sc));
 			state.world.province_set_is_slave(id, state.world.province_get_is_slave(sc));
-			if(state.world.province_get_building_level(id, economy::province_building_type::naval_base) > 0) {
+			if(state.world.province_get_building_level(id, uint8_t(economy::province_building_type::naval_base)) > 0) {
 				if(state.world.state_instance_get_naval_base_is_taken(new_si)) {
-					state.world.province_set_building_level(id, economy::province_building_type::naval_base, 0);
+					state.world.province_set_building_level(id, uint8_t(economy::province_building_type::naval_base), 0);
 				} else {
 					state.world.state_instance_set_naval_base_is_taken(new_si, true);
 				}
@@ -778,7 +778,7 @@ void change_province_owner(sys::state& state, dcon::province_id id, dcon::nation
 	} else {
 		state.world.province_set_state_membership(id, dcon::state_instance_id{});
 		for(auto t = economy::province_building_type::railroad; t != economy::province_building_type::last; t = economy::province_building_type(uint8_t(t) + 1)) {
-			state.world.province_set_building_level(id, t, uint8_t(0));
+			state.world.province_set_building_level(id, uint8_t(t), uint8_t(0));
 		}
 
 		auto province_fac_range = state.world.province_get_factory_location(id);
@@ -1196,7 +1196,7 @@ bool can_start_colony(sys::state& state, dcon::nation_id n, dcon::state_definiti
 
 	if(!adjacent && coastal_target  && state.world.nation_get_central_ports(n) != 0) {
 		for(auto p : state.world.nation_get_province_ownership(n)) {
-			if(auto nb_level = p.get_province().get_building_level(economy::province_building_type::naval_base); nb_level > 0 && p.get_province().get_nation_from_province_control() == n) {
+			if(auto nb_level = p.get_province().get_building_level(uint8_t(economy::province_building_type::naval_base)); nb_level > 0 && p.get_province().get_nation_from_province_control() == n) {
 				auto dist = province::direct_distance(state, p.get_province(), coastal_target);
 				if(dist <= province::world_circumference *  0.04f * nb_level) {
 					reachable_by_sea = true;
@@ -1298,7 +1298,7 @@ bool fast_can_start_colony(sys::state& state, dcon::nation_id n, dcon::state_def
 
 	if(!adjacent && coastal_target && state.world.nation_get_central_ports(n) != 0) {
 		for(auto p : state.world.nation_get_province_ownership(n)) {
-			if(auto nb_level = p.get_province().get_building_level(economy::province_building_type::naval_base); nb_level > 0 && p.get_province().get_nation_from_province_control() == n) {
+			if(auto nb_level = p.get_province().get_building_level(uint8_t(economy::province_building_type::naval_base)); nb_level > 0 && p.get_province().get_nation_from_province_control() == n) {
 				auto dist = province::direct_distance(state, p.get_province(), coastal_target);
 				if(dist <= province::world_circumference * 0.04f * nb_level) {
 					reachable_by_sea = true;
